Move utils.go to utils directory
