+++ evilwallet/evilwallet.go
@@ -8,6 +8,7 @@ import (
 	"github.com/iotaledger/evil-tools/accountwallet"
 	evillogger "github.com/iotaledger/evil-tools/logger"
 	"github.com/iotaledger/evil-tools/models"
+	"github.com/iotaledger/evil-tools/utils"
 	"github.com/iotaledger/hive.go/ds/types"
 	"github.com/iotaledger/hive.go/ierrors"
 	"github.com/iotaledger/hive.go/lo"
@@ -285,7 +286,7 @@ func (e *EvilWallet) requestFaucetFunds(wallet *Wallet) (output *models.Output, 
 		return nil, ierrors.Wrap(err, "failed to request funds from faucet")
 	}
 
-	outputID, iotaOutput, err := AwaitAddressUnspentOutputToBeAccepted(clt, receiveAddr, 10*time.Second)
+	outputID, iotaOutput, err := utils.AwaitAddressUnspentOutputToBeAccepted(clt, receiveAddr, 10*time.Second)
 	if err != nil {
 		return nil, ierrors.Wrap(err, "failed to await faucet output acceptance")
 	}
@@ -371,7 +372,7 @@ func (e *EvilWallet) splitOutputs(inputWallet, outputWallet *Wallet) ([]iotago.T
 }
 
 func (e *EvilWallet) createSplitOutputs(input *models.Output, splitNumber int, receiveWallet *Wallet) []*OutputOption {
-	balances := SplitBalanceEqually(splitNumber, input.Balance)
+	balances := utils.SplitBalanceEqually(splitNumber, input.Balance)
 	outputs := make([]*OutputOption, splitNumber)
 	for i, bal := range balances {
 		outputs[i] = &OutputOption{amount: bal, address: receiveWallet.Address(), outputType: iotago.OutputBasic}
@@ -744,7 +745,7 @@ func (e *EvilWallet) updateOutputBalances(buildOptions *Options) (err error) {
 				totalBalance += in.Balance
 			}
 		}
-		balances := SplitBalanceEqually(len(buildOptions.outputs)+len(buildOptions.aliasOutputs), totalBalance)
+		balances := utils.SplitBalanceEqually(len(buildOptions.outputs)+len(buildOptions.aliasOutputs), totalBalance)
 		i := 0
 		for out, output := range buildOptions.aliasOutputs {
 			switch output.Type() {
